+++ app/src/main/java/com/omelan/cofi/pages/settings/licenses/LicensesList.kt
@@ -10,13 +10,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import com.omelan.cofi.R
 import com.omelan.cofi.components.PiPAwareAppBar
 import com.omelan.cofi.components.createAppBarBehavior
-import com.omelan.cofi.ui.Spacing
+import com.omelan.cofi.utils.getDefaultPadding
 import com.omelan.cofi.utils.parseJsonToDependencyList
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -24,8 +23,6 @@ import com.omelan.cofi.utils.parseJsonToDependencyList
 fun LicensesList(goBack: () -> Unit) {
     val context = LocalContext.current
     val appBarBehavior = createAppBarBehavior()
-    val navigationBarPadding = WindowInsets.navigationBars.asPaddingValues()
-    val layoutDirection = LocalLayoutDirection.current
 
     val dependencyList = context.assets.open("open_source_licenses.json").bufferedReader().use {
         it.readText()
@@ -53,16 +50,7 @@ fun LicensesList(goBack: () -> Unit) {
             modifier = Modifier
                 .nestedScroll(appBarBehavior.nestedScrollConnection)
                 .fillMaxSize(),
-            contentPadding = PaddingValues(
-                start = navigationBarPadding.calculateStartPadding(layoutDirection) +
-                    it.calculateStartPadding(layoutDirection) + Spacing.big,
-                top = navigationBarPadding.calculateTopPadding() +
-                    it.calculateTopPadding() + Spacing.small,
-                bottom = navigationBarPadding.calculateBottomPadding() +
-                    it.calculateBottomPadding() + Spacing.big,
-                end = navigationBarPadding.calculateEndPadding(layoutDirection) +
-                    it.calculateEndPadding(layoutDirection) + Spacing.big
-            ),
+            contentPadding = getDefaultPadding(paddingValues = it),
         ) {
             items(dependencyList) {
                 DependencyItem(dependency = it)
